testing

+++ includes/src/Extensions/Config/Configurator.php
@@ -2,10 +2,12 @@
 
 namespace JTL\Extensions\Config;
 
+use JTL\Cart\Cart;
 use JTL\Cart\CartHelper;
 use JTL\Catalog\Product\Artikel;
 use JTL\Catalog\Product\Preise;
 use JTL\Nice;
+use JTL\Session\Frontend;
 use JTL\Shop;
 use function Functional\some;
 
@@ -111,9 +113,9 @@ class Configurator
     }
 
     /**
-     * @param object $cart
+     * @param Cart $cart
      */
-    public static function postcheckCart($cart): void
+    public static function postcheckCart(Cart $cart): void
     {
         if (!\is_array($cart->PositionenArr) || \count($cart->PositionenArr) === 0 || !self::checkLicense()) {
             return;
@@ -124,7 +126,7 @@ class Configurator
                 continue;
             }
             $deleted = false;
-            if ($item->cUnique && (int)$item->kKonfigitem === 0) {
+            if ($item->cUnique && $item->kKonfigitem === 0) {
                 $configItems = [];
                 foreach ($cart->PositionenArr as $child) {
                     if ($child->cUnique && $child->cUnique === $item->cUnique && $child->kKonfigitem > 0) {
@@ -138,7 +140,7 @@ class Configurator
                 }
             } elseif (!$item->cUnique) {
                 // Konfiguration vorhanden -> löschen
-                if (self::hasKonfig($item->kArtikel)) {
+                if ($item->kKonfigitem > 0 && self::hasKonfig($item->kArtikel)) {
                     $deleted        = true;
                     $deletedItems[] = $index;
                 }
@@ -150,7 +152,10 @@ class Configurator
                 );
             }
         }
-        CartHelper::deleteCartItems($deletedItems, false);
+        Shop::dbg(Frontend::getCustomer()->getID(), false, 'kkunde:');
+        if (\count($deletedItems) > 0) {
+            CartHelper::deleteCartItems($deletedItems, false);
+        }
     }
 
     /**
